pyproject.toml
[tool.poetry]
name = "pptb"
version = "0.1.9-alpha.1"
description = "🚣 一些常用的但 paddle 里没有的小工具～"
authors = ["Nyakku Shigure <sigure.qaq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/cattidea/paddle-toolbox"
repository = "https://github.com/cattidea/paddle-toolbox"
keywords = ["python", "paddle", "toolbox"]

[tool.poetry.dependencies]
python = "^3.7"
paddlepaddle = { version = "^2.3.0rc0", optional = true, allow-prereleases = true }

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^7.1.2"
isort = "^5.10.1"

[tool.poetry.extras]
paddle = ["paddlepaddle"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/cattidea/paddle-toolbox/issues"

[tool.pytest.ini_options]
markers = ["ci_skip", "ci_only"]
filterwarnings = ["ignore:Skip loading for", "ignore:When training"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
